SkAndroidCode: Remove most of CICP copy
This is a duplicate of the code in SkColorSpace, so it should be
deleted.

Mostly. The HLG and PQ transfer functions are slightly different, and
that caused this to be reverted here:
https://skia-review.googlesource.com/c/skia/+/1048476

While in the neighborhood, change the SkNamedPrimaries::GetCicp function
to populate an skcms_Matrix3x3, because that's what all of the callers
want anyway (and it allows to keep the output of SkAndroidCodec
unchanged).

This is towards moving all of this into SkCodecs::ColorProfile, so that
we no longer access the skcms_ICCProfile directly.

Bug: skia:456675139
Change-Id: I4e406bc758ccec0ff0435a1f9afe283c5483acef
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1112939
Reviewed-by: Christopher Cameron <ccameron@google.com>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>