SkCodec: Remove some accessors to skcms_ICCProfile
This is towards removing all accessors and hiding the use of skcms for
parsing and for transform behind a single structure.

Several codecs check if the ICC profile is RGB, Gray, or CMYK. Add an
accessor for that.

Add accessors for the various ways of creating an SkColorSpace from an
ICC profile (the SkColorSpace::Make version and the slightly-different
Android version).

Bug: skia:456675139
Change-Id: Ic73907987352ca4acfbfcbaaa7eebe5dcc839447
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1111961
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Christopher Cameron <ccameron@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>