[wgsl] Refactor assemblePartialSampleCall to control closing parens
Also updates handling of the `textureLoad` to just use
`assemblePartialSampleCall` with some extra logic for
`assemblePartialSampleCall` to know whether or not it needs to emit just
a texture or both texture and sampler.

This will be helpful in the future when adding automatic casts to f16
for sample and texture calls.

Bug: b/465408252
Change-Id: I831cefe8b1fdd3673e99afacd1f00a6ad902cdc4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1118496
Reviewed-by: Thomas Smith <thomsmit@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>