[graphite] Use largest supported sample count < max internal sample count
This changes the behavior if the client requested a large
not-always-available sample count. Before, this would cause Device to
fall back all the way to using the raster atlas. Now it'll select the
best quality supported on the format that is <= the requested limit. The
motivation for this behavior change is for more consistency across
formats in the path strategy they use. Since 4x MSAA is pretty much
supported everywhere for renderable formats, this means there will
always be some MSAA unless disabled by the context options. When the
strategy choice is moved to the RendererProvider, it won't be changing
behavior after this change.

This also bubbles up the concept of "avoidMSAA" from a HW
capabilities/performance POV that is detected by Caps. This will be used
in follow up work to help choose the explicit path strategy.

Bug: b/452351267
Change-Id: I699f4a70b812e86451dabdc2ce460af7bbb1df18
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1100396
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>