[wgsl] Stop forcing all function calls into scratch lets
This leads to decent reduction in the size of the generated WGSL and the
WGSL ends up matching closer to the GLSL or MSL that the other code
generators emit.

We preserve writing to a let or to a statement when handling out
parameters since there's a mismatch in SkSL's semantics vs. WGSL's.
However, this logic is now consolidated into assembleFunctionCall()
instead of spread out across that call, assembleIntrinsicCall(), and
assembleSimpleIntrinsic(), etc.

This is also part of a larger effort to pull back the use of
writeScratchLet(), which poses difficulties with handling f16 vs f32.

Bug: b/465408252
Change-Id: I8c20cd9544c686a2578316d54fc85462e92a0f7c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1118497
Reviewed-by: Thomas Smith <thomsmit@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>