Enable user-defined memory limits for decoding jpeg
This tracks a large common allocation in SkCodec (for all codecs)
and the (relatively small) allocation in Skia. It estimates
what libjpeg-turbo will use before trying to read the pixel.

If we like this approach, we can implement it for more codecs
(some may be more accurate than others).

Change-Id: Ic16b0ccd9711615b1983555f5f4f63b1b2c905f3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1106781
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>