[pdf] Fix typeface duplication due to luminance
Ensure strikes are created with SkScalerContextFlags::kNone (in
particular without SkScalerContextFlags::kFakeGamma) to indicate that
fake gamma will not be used with these strikes. When fake gamma is
enabled the luminance of the current color will be used as part of the
strike so that the fake gamma can be computed, but this is not needed or
wanted when emitting PDF. Enabling fake gamma leads to typefaces being
embedded in the PDF more than once if the typeface is used for different
colors as the strikes are different even though they logically are not.

Change-Id: Ic8467f950bd2193534b2f3a86b7dfd93c5af2617
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1132857
Auto-Submit: Ben Wagner <bungeman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>